fixed starting issues
