--- index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Tic-tac-toe</title>
-  <link href="style.css" rel="stylesheet" type="text/css" />
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Dhurjati&display=swap" rel="stylesheet">
-</head>
-<body id="body">
-  <div id="general-wrapper">
-    <div id="start-content">
-      <p id="welcome-text">enter your names</p>
-      <p>first player</p>
-      <input class="name-input" id="first-input" />
-      <p>second player</p>
-      <input class="name-input" id="second-input" />
-      <button id="confirm-button">confirm</button>
-    </div>
-  </div>
-  <script type="module" src="main.js"></script>
-  </body>
-</html>
