+++ package.json
@@ -1,14 +1,25 @@
 {
-  "name": "tic-tac-toe",
+  "name": "simple-webpack",
   "version": "1.0.0",
-  "description": "Tic Tac Toe application",
-  "main": "main.js",
+  "description": "",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "rimraf ./dist && webpack",
+    "start": "webpack-dev-server",
+    "prettier": "prettier --write ../example-simple-webpack"
   },
+  "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "webpack": "^5.89.0"
+    "css-loader": "^6.8.1",
+    "html-webpack-plugin": "^5.5.3",
+    "prettier": "^2.8.8",
+    "rimraf": "^5.0.5",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "html-loader": "^4.2.0"
   }
 }
