+++ src/index.js
@@ -1,5 +1,7 @@
-import {checking} from "/check.js";
-import {myTurn, winner} from "/myTurn.js";
+import './styles.css';
+import backgroundUrl from './images/board_background.jpg';
+import noughtUrl from './images/nought.jpg';
+import crossUrl from './images/cross.jpg';
 
 const body = document.querySelector("#body");
 const generalWrapper = document.querySelector("#general-wrapper");
@@ -50,7 +52,7 @@ if (confirmButton) (
 )
 
 function game() {
-    generalWrapper.style = 'background-image: url("/graphic/board_background.jpg")';
+    generalWrapper.style = `background-image: url(${backgroundUrl})`;
     scoreBox.id = "score-box";
     body.append(scoreBox);
     firstPlayerName.classList.add("score", "player-name");
@@ -98,13 +100,13 @@ function game() {
     spaceOne.addEventListener("click", function(){
         if (score[1] === 0) {
             if (turn === 1) {
-                spaceOne.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceOne.style = `background-image: url(${noughtUrl})`;
                 score[1] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceOne.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceOne.style = `background-image: url(${crossUrl})`;
                 score[1] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -115,13 +117,13 @@ function game() {
     spaceTwo.addEventListener("click", function(){
         if (score[2] === 0) {
             if (turn === 1) {
-                spaceTwo.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceTwo.style = `background-image: url(${noughtUrl})`;
                 score[2] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceTwo.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceTwo.style = `background-image: url(${crossUrl})`;
                 score[2] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -132,13 +134,13 @@ function game() {
     spaceThree.addEventListener("click", function(){
         if (score[3] === 0) {
             if (turn === 1) {
-                spaceThree.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceThree.style = `background-image: url(${noughtUrl})`;
                 score[3] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceThree.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceThree.style = `background-image: url(${crossUrl})`;
                 score[3] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -149,13 +151,13 @@ function game() {
     spaceFour.addEventListener("click", function(){
         if (score[4] === 0) {
             if (turn === 1) {
-                spaceFour.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceFour.style = `background-image: url(${noughtUrl})`;
                 score[4] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceFour.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceFour.style = `background-image: url(${crossUrl})`;
                 score[4] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -166,13 +168,13 @@ function game() {
     spaceFive.addEventListener("click", function(){
         if (score[5] === 0) {
             if (turn === 1) {
-                spaceFive.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceFive.style = `background-image: url(${noughtUrl})`;
                 score[5] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceFive.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceFive.style = `background-image: url(${crossUrl})`;
                 score[5] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -183,13 +185,13 @@ function game() {
     spaceSix.addEventListener("click", function(){
         if (score[6] === 0) {
             if (turn === 1) {
-                spaceSix.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceSix.style = `background-image: url(${noughtUrl})`;
                 score[6] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceSix.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceSix.style = `background-image: url(${crossUrl})`;
                 score[6] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -200,13 +202,13 @@ function game() {
     spaceSeven.addEventListener("click", function(){
         if (score[7] === 0) {
             if (turn === 1) {
-                spaceSeven.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceSeven.style = `background-image: url(${noughtUrl})`;
                 score[7] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceSeven.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceSeven.style = `background-image: url(${crossUrl})`;
                 score[7] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -217,13 +219,13 @@ function game() {
     spaceEight.addEventListener("click", function(){
         if (score[8] === 0) {
             if (turn === 1) {
-                spaceEight.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceEight.style = `background-image: url(${noughtUrl})`;
                 score[8] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceEight.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceEight.style = `background-image: url(${crossUrl})`;
                 score[8] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -234,13 +236,13 @@ function game() {
     spaceNine.addEventListener("click", function(){
         if (score[9] === 0) {
             if (turn === 1) {
-                spaceNine.style = 'background-image: url("/graphic/nought.jpg")';
+                spaceNine.style = `background-image: url(${noughtUrl})`;
                 score[9] = 1;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 2;
                 myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
             } else {
-                spaceNine.style = 'background-image: url("/graphic/cross.jpg")';
+                spaceNine.style = `background-image: url(${crossUrl})`;
                 score[9] = 2;
                 winner(checking(score), isFirstPlayerTurn, isSecondPlayerTurn, drawLine);
                 turn = 1;
@@ -249,4 +251,74 @@ function game() {
         }
     })
     myTurn(turn, isFirstPlayerTurn, isSecondPlayerTurn);
+}
+
+function checking(object) {
+    if (object[1] === object[2] && object[2] === object[3] && object[1] !== 0) {
+        return object[1];
+    }
+    if (object[1] === object[5] && object[5] === object[9] && object[1] !== 0) {
+        return object[1];
+    }
+    if (object[1] === object[4] && object[4] === object[7] && object[1] !== 0) {
+        return object[1];
+    }
+    if (object[5] === object[2] && object[2] === object[8] && object[5] !== 0) {
+        return object[5];
+    }
+    if (object[6] === object[9] && object[9] === object[3] && object[3] !== 0) {
+        return object[6];
+    }
+    if (object[4] === object[5] && object[5] === object[6] && object[5] !== 0) {
+        return object[4];
+    }
+    if (object[3] === object[7] && object[7] === object[5] && object[3] !== 0) {
+        return object[3];
+    }
+    if (object[7] === object[8] && object[8] === object[9] && object[8] !== 0) {
+        return object[7];
+    }
+    if (object[1] !== 0 && object[2] !== 0 && object[3] !== 0 && object[4] !== 0 && object[5] !== 0 && object[6] !== 0 && object[7] !== 0 && object[8] !== 0 && object[9] !== 0) {
+        return 3;
+    }
+}
+
+function myTurn(turn, firstPlayer, secondPlayer) {
+    if (turn === 1) {
+        firstPlayer.innerHTML = "your turn!";
+        secondPlayer.innerHTML = "";
+    } else {
+        secondPlayer.innerHTML = "your turn!";
+        firstPlayer.innerHTML = "";
+    }
+}
+
+function winner(outcome, firstPlayer, secondPlayer, drawLine) {
+    if (outcome === 1) {
+        console.log("one really won");
+        firstPlayer.innerHTML = "won!";
+        firstPlayer.style = "color: red; font-size: 1.5em";
+        setTimeout(function() {
+            location.reload();
+        }, 3000);
+        throw new Error();
+    }
+    if (outcome === 2) {
+        secondPlayer.innerHTML = "won!";
+        secondPlayer.style = "color: red; font-size: 1.5em";
+        setTimeout(function() {
+            location.reload();
+        }, 3000);
+        throw new Error();
+    }
+    if (outcome === 3) {
+        drawLine.innerHTML = "draw!";
+        secondPlayer.innerHTML = "";
+        firstPlayer.innerHTML = "";
+        drawLine.style = "color: red; font-size: 1.5em";
+        setTimeout(function() {
+            location.reload();
+        }, 3000);
+        throw new Error();
+    }
 }
