--- style.css
@@ -1,105 +1,0 @@
-* {
-    box-sizing: border-box;
-    text-transform: uppercase;
-    font-family: 'Dhurjati', sans-serif;
-    font-size: 30px;
-    margin: 0;
-
-    @media (max-width: 600px) {
-        font-size: 5vw;
-    }
-
-}
-
-#body {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-
-    @media (max-width: 900px) {
-        flex-direction: column;
-        align-items: center;
-    }
-}
-
-#general-wrapper {
-    width: 600px;
-    height: 600px;
-    margin-top: 5px;
-    background-size: cover;
-    display: flex;
-    flex-wrap: wrap;
-    row-gap: 0;
-
-    @media (max-width: 600px) {
-        width:100%;
-        height: 100vw;
-    }
-}
-
-#start-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-#welcome-text {
-    font-size: 2.5em;
-}
-
-#confirm-button {
-    margin-top: 55px;
-
-    @media (max-width: 400px) {
-        margin-top: 3vh;
-    }
-}
-
-#score-box {
-    width: 300px;
-    height: 300px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-
-    @media (max-width: 900px) {
-        width: 600px;
-        height: 300px;
-    }
-    @media (max-width: 600px) {
-        width:100%;
-        height: 100vw;
-    }
-}
-
-#player-to-start {
-    text-align: center;
-    font-size: 2.5em;
-    padding-top: 75px;
-}
-
-.space {
-    width: 190px;
-    height: 190px;
-    background-color: white;
-    background-size: cover;
-    border: none;
-    margin: 5px;
-
-    @media (max-width: 600px) {
-        width: 31.66vw;
-        height: 31.66vw;
-        margin: 0.83vw;
-    }
-}
-
-.score {
-    margin: 0;
-    padding-left: 30px;
-    height: 40px;
-}
-
-.player-name {
-    font-size: 1.5em;
-}
